refactor: driver: Move MongoRepository to driver
