+++ src/entity-manager/EntityManager.ts
@@ -9,7 +9,7 @@ import {DeepPartial} from "../common/DeepPartial";
 import {RemoveOptions} from "../repository/RemoveOptions";
 import {SaveOptions} from "../repository/SaveOptions";
 import {NoNeedToReleaseEntityManagerError} from "../error/NoNeedToReleaseEntityManagerError";
-import {MongoRepository} from "../repository/MongoRepository";
+import {MongoRepository} from "../driver/mongodb/MongoRepository";
 import {TreeRepository} from "../repository/TreeRepository";
 import {Repository} from "../repository/Repository";
 import {FindOptionsUtils} from "../find-options/FindOptionsUtils";
