+++ src/index.ts
@@ -11,7 +11,7 @@ import {Repository} from "./repository/Repository";
 import {EntityManager} from "./entity-manager/EntityManager";
 import {PlatformTools} from "./platform/PlatformTools";
 import {TreeRepository} from "./repository/TreeRepository";
-import {MongoRepository} from "./repository/MongoRepository";
+import {MongoRepository} from "./driver/mongodb/MongoRepository";
 import {ConnectionOptionsReader} from "./connection/ConnectionOptionsReader";
 import {MongoEntityManager} from "./driver/mongodb/MongoEntityManager";
 import {SqljsEntityManager} from "./driver/sqljs/SqljsEntityManager";
@@ -109,7 +109,7 @@ export * from "./repository/AbstractRepository";
 export * from "./repository/Repository";
 export * from "./repository/BaseEntity";
 export * from "./repository/TreeRepository";
-export * from "./repository/MongoRepository";
+export * from "./driver/mongodb/MongoRepository";
 export * from "./repository/RemoveOptions";
 export * from "./repository/SaveOptions";
 export * from "./schema-builder/table/TableCheck";
